+++ R/theme.R
@@ -3,9 +3,8 @@
 ##'
 ##' @title theme_tree
 ##' @param bgcolor background color
-##' @param fgcolor foreground color
 ##' @param ... additional parameter
-##' @importFrom ggplot2 theme_bw
+##' @importFrom ggplot2 theme_void
 ##' @importFrom ggplot2 theme
 ##' @importFrom ggplot2 element_blank
 ##' @importFrom ggplot2 xlab
@@ -17,16 +16,21 @@
 ##' require(ape)
 ##' tr <- rtree(10)
 ##' ggtree(tr) + theme_tree()
-theme_tree <- function(bgcolor="white", fgcolor="black", ...) {
-    list(xlab(NULL),
-         ylab(NULL),
-	 theme_tree2_internal() +
-         theme(panel.background=element_rect(fill=bgcolor, colour=bgcolor),
-               axis.line.x = element_blank(),
-               axis.text.x = element_blank(),
-               axis.ticks.x = element_blank(),
-               ...)
-	 )
+theme_tree <- function(bgcolor="white", ...) {
+
+    ## list(xlab(NULL),
+    ##      ylab(NULL),
+    ##      theme_tree2_internal() +
+    ##      theme(panel.background=element_rect(fill=bgcolor, colour=bgcolor),
+    ##            axis.line.x = element_blank(),
+    ##            axis.text.x = element_blank(),
+    ##            axis.ticks.x = element_blank(),
+    ##            ...)
+    ##      )
+
+    theme_void() +
+    theme(panel.background=element_rect(fill=bgcolor, colour=bgcolor),
+          ...)
 }
 
 ##' dendrogram theme
@@ -34,15 +38,19 @@ theme_tree <- function(bgcolor="white", fgcolor="black", ...) {
 ##'
 ##' @title theme_dendrogram
 ##' @inheritParams theme_tree
+##' @param fgcolor color for axis
 ##' @export
 ##' @importFrom ggplot2 element_text
 ##' @author Guangchuang Yu
-theme_dendrogram <- function(bgcolor="white", fgcolor="black", ...) {
-    theme_tree(bgcolor = bgcolor, fgcolor = fgcolor,
-               axis.line.y = element_line(color=fgcolor),
-               axis.text.y = element_text(color=fgcolor),
-               axis.ticks.y = element_line(color=fgcolor),
-               ...)
+theme_dendrogram <- function(bgcolor = "white", fgcolor = "black", ...) {
+    theme_tree2(bgcolor = bgcolor,
+                axis.line.x = element_blank(),
+                axis.text.x = element_blank(),
+                axis.ticks.x = element_blank(),
+                axis.line.y = element_line(color=fgcolor),
+                axis.text.y = element_text(color=fgcolor),
+                axis.ticks.y = element_line(color=fgcolor),
+                ...)
 }
 
 
@@ -72,7 +80,7 @@ theme_tree2 <- function(bgcolor="white", fgcolor="black", ...) {
 	  )
 }
 
-##' @importFrom ggplot2 theme_void
+##' @importFrom ggplot2 theme_bw
 theme_tree2_internal <- function(bgcolor="white", fgcolor="black",
                                  legend.position="right",
                                  panel.grid.minor=element_blank(),
@@ -83,7 +91,8 @@ theme_tree2_internal <- function(bgcolor="white", fgcolor="black",
                                  axis.text.y=element_blank(),...) {
     ## need to set axis.line otherwise the setting cannot be inherited.
     ## https://github.com/GuangchuangYu/ggtree/issues/218
-    theme_void() +
+
+    theme_bw() +
         theme(legend.position=legend.position,
               panel.grid.minor=panel.grid.minor,
               panel.grid.major=panel.grid.major,